# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Mora e-commerce admin API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/analytics/summary:
    get:
      operationId: adminGetAnalyticsSummary
      tags: [analytics]
      security: [{ bearerAuth: [] }]
      summary: Get analytics overview
      responses:
        "200":
          description: Analytics summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsSummaryResponse"
  /admin/analytics/revenue:
    get:
      operationId: adminGetRevenueChart
      tags: [analytics]
      security: [{ bearerAuth: [] }]
      summary: Revenue over time
      parameters:
        - name: days
          in: query
          schema:
            type: integer
            default: 14
      responses:
        "200":
          description: Revenue chart data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RevenueChartResponse"
  /admin/analytics/top-products:
    get:
      operationId: adminGetTopProducts
      tags: [analytics]
      security: [{ bearerAuth: [] }]
      summary: Top selling products
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 5
      responses:
        "200":
          description: Top products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TopProductsResponse"
  /admin/analytics/reports:
    get:
      operationId: adminGetAnalyticsReports
      tags: [analytics]
      security: [{ bearerAuth: [] }]
      summary: Key metric reports
      responses:
        "200":
          description: Reports
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsReportsResponse"
...
  /admin/customers/segments:
    get:
      operationId: adminListCustomerSegments
      tags: [customers]
      security: [{ bearerAuth: [] }]
      summary: List customer segments
      responses:
        "200":
          description: Segments
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerSegmentsResponse"
  /admin/customers/companies:
    get:
      operationId: adminListCustomerCompanies
      tags: [customers]
      security: [{ bearerAuth: [] }]
      summary: List B2B companies
      responses:
        "200":
          description: Companies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerCompaniesResponse"
  /admin/customers:
    get:
      operationId: adminListCustomers
      tags: [customers]
      security: [{ bearerAuth: [] }]
      summary: List customers
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: segment
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Customer list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerListResponse"
    post:
      operationId: adminCreateCustomer
      tags: [customers]
      security: [{ bearerAuth: [] }]
      summary: Create a customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerInput"
      responses:
        "201":
          description: Created customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerResponse"
  /admin/customers/{id}:
    get:
      operationId: adminGetCustomer
      tags: [customers]
      security: [{ bearerAuth: [] }]
      summary: Get a customer with orders
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Customer detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerDetailResponse"
    put:
      operationId: adminUpdateCustomer
      tags: [customers]
      security: [{ bearerAuth: [] }]
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CustomerUpdate"
      responses:
        "200":
          description: Updated customer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerResponse"
    delete:
      operationId: adminDeleteCustomer
      tags: [customers]
      security: [{ bearerAuth: [] }]
      summary: Delete a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeletedResponse"

  # ─── Campaigns ───────────────────────────────────────────────────────────────
...
  /admin/markets:
    get:
      operationId: adminListMarkets
      tags: [markets]
      security: [{ bearerAuth: [] }]
      summary: List markets
      responses:
        "200":
          description: Market list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketListResponse"
    post:
      operationId: adminCreateMarket
      tags: [markets]
      security: [{ bearerAuth: [] }]
      summary: Create a market
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarketInput"
      responses:
        "201":
          description: Created market
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketResponse"
  /admin/markets/{id}:
    get:
      operationId: adminGetMarket
      tags: [markets]
      security: [{ bearerAuth: [] }]
      summary: Get a market
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Market
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketResponse"
    put:
      operationId: adminUpdateMarket
      tags: [markets]
      security: [{ bearerAuth: [] }]
      summary: Update a market
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarketUpdate"
      responses:
        "200":
          description: Updated market
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketResponse"
    delete:
      operationId: adminDeleteMarket
      tags: [markets]
      security: [{ bearerAuth: [] }]
      summary: Delete a market
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeletedResponse"

  # ─── Settings ─────────────────────────────────────────────────────────────────
  /admin/settings:
    get:
      operationId: adminGetSettings
      tags: [settings]
      security: [{ bearerAuth: [] }]
      summary: Get store settings
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettingsResponse"
    put:
      operationId: adminUpdateSettings
      tags: [settings]
      security: [{ bearerAuth: [] }]
      summary: Update store settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingsUpdate"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettingsResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    DeletedResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            deleted:
              type: boolean
          required: [deleted]
        meta:
          type: object
        error:
          type: ["string", "null"]
      required: [data, meta, error]
    ListMeta:
      type: object
      properties:
        total:
          type: integer
      required: [total]

    # ─── Product ──────────────────────────────────────────────────────────────
...
    Order:
      type: object
      properties:
        id: { type: string }
        orderNumber: { type: string }
        customerId: { type: ["string", "null"] }
        email: { type: string }
        status: { type: string }
        financialStatus: { type: string }
        fulfillmentStatus: { type: string }
        subtotal: { type: number }
        shipping: { type: number }
        tax: { type: number }
        total: { type: number }
        currency: { type: string }
        shippingAddress: { type: object }
        lineItems: { type: array, items: { type: object } }
        note: { type: string }
        tags: { type: array, items: { type: string } }
        isDraft: { type: boolean }
        isAbandoned: { type: boolean }
        createdAt: { type: string }
        updatedAt: { type: string }
      required: [id, orderNumber, email, status, total]
...
    Customer:
      type: object
      properties:
        id: { type: string }
        firstName: { type: string }
        lastName: { type: string }
        email: { type: string }
        phone: { type: string }
        ordersCount: { type: integer }
        totalSpent: { type: number }
        tags: { type: array, items: { type: string } }
        segment: { type: ["string", "null"] }
        company: { type: ["string", "null"] }
        address: { type: object }
        acceptsMarketing: { type: boolean }
        createdAt: { type: string }
      required: [id, firstName, lastName, email]
    CustomerDetail:
      allOf:
        - $ref: "#/components/schemas/Customer"
        - type: object
          properties:
            orders:
              type: array
              items:
                $ref: "#/components/schemas/Order"
          required: [orders]
    CustomerSegment:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        count: { type: integer }
      required: [id, name, count]
    CustomerCompany:
      type: object
      properties:
        company: { type: string }
        customerCount: { type: integer }
      required: [company, customerCount]
    CustomerInput:
      type: object
      properties:
        firstName: { type: string }
        lastName: { type: string }
        email: { type: string }
        phone: { type: string }
        address: { type: object }
        acceptsMarketing: { type: boolean }
      required: [firstName, lastName, email]
    CustomerUpdate:
      type: object
      properties:
        firstName: { type: string }
        lastName: { type: string }
        email: { type: string }
        phone: { type: string }
        segment: { type: string }
        company: { type: string }
    CustomerListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Customer"
        meta:
          $ref: "#/components/schemas/ListMeta"
        error:
          type: ["string", "null"]
      required: [data, meta, error]
    CustomerResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Customer"
        meta:
          type: object
        error:
          type: ["string", "null"]
      required: [data, meta, error]
    CustomerDetailResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/CustomerDetail"
        meta:
          type: object
        error:
          type: ["string", "null"]
      required: [data, meta, error]
    CustomerSegmentsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/CustomerSegment"
        meta:
          $ref: "#/components/schemas/ListMeta"
        error:
          type: ["string", "null"]
      required: [data, meta, error]
    CustomerCompaniesResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/CustomerCompany"
        meta:
          $ref: "#/components/schemas/ListMeta"
        error:
          type: ["string", "null"]
      required: [data, meta, error]

    # ─── Campaign ─────────────────────────────────────────────────────────────
...
    Market:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        countries: { type: array, items: { type: string } }
        currency: { type: string }
        status: { type: string }
        createdAt: { type: string }
      required: [id, name, currency, status]
    MarketInput:
      type: object
      properties:
        name: { type: string }
        countries: { type: array, items: { type: string } }
        currency: { type: string }
      required: [name, currency]
    MarketUpdate:
      type: object
      properties:
        name: { type: string }
        currency: { type: string }
        status: { type: string }
    MarketListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Market"
        meta:
          $ref: "#/components/schemas/ListMeta"
        error:
          type: ["string", "null"]
      required: [data, meta, error]
    MarketResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Market"
        meta:
          type: object
        error:
          type: ["string", "null"]
      required: [data, meta, error]

    # ─── Settings ─────────────────────────────────────────────────────────────
    StoreSettings:
      type: object
      properties:
        storeName: { type: string }
        storeEmail: { type: string }
        storePhone: { type: string }
        storeAddress: { type: string }
        currency: { type: string }
        timezone: { type: string }
        language: { type: string }
    SettingsUpdate:
      type: object
      properties:
        storeName: { type: string }
        storeEmail: { type: string }
        storePhone: { type: string }
        storeAddress: { type: string }
        currency: { type: string }
        timezone: { type: string }
        language: { type: string }
    SettingsResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/StoreSettings"
        meta:
          type: object
        error:
          type: ["string", "null"]
      required: [data, meta, error]

    # ─── Analytics ────────────────────────────────────────────────────────────
    AnalyticsSummary:
      type: object
      properties:
        revenue: { type: number }
        orders: { type: integer }
        customers: { type: integer }
        avgOrderValue: { type: number }
        products: { type: integer }
      required: [revenue, orders, customers, avgOrderValue, products]
    RevenuePoint:
      type: object
      properties:
        date: { type: string }
        revenue: { type: number }
        orders: { type: integer }
      required: [date, revenue, orders]
    TopProduct:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        revenue: { type: number }
        unitsSold: { type: integer }
      required: [id, title, revenue, unitsSold]
    AnalyticsReport:
      type: object
      properties:
        name: { type: string }
        value: {}
        change: { type: string }
      required: [name, value, change]
    AnalyticsSummaryResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/AnalyticsSummary"
        meta:
          type: object
        error:
          type: ["string", "null"]
      required: [data, meta, error]
    RevenueChartResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/RevenuePoint"
        meta:
          type: object
        error:
          type: ["string", "null"]
      required: [data, meta, error]
    TopProductsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/TopProduct"
        meta:
          type: object
        error:
          type: ["string", "null"]
      required: [data, meta, error]
    AnalyticsReportsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AnalyticsReport"
        meta:
          type: object
        error:
          type: ["string", "null"]
      required: [data, meta, error]